+++ packages/file-type-icons/src/FileIconType.ts
@@ -11,7 +11,8 @@ export enum FileIconType {
   genericFile = 3,
   listItem = 4,
   sharedFolder = 5,
-  multiple = 6
+  multiple = 6,
+  stream = 7
 }
 
-export type FileIconTypeInput = 1 | 2 | 3 | 4 | 5 | 6;
+export type FileIconTypeInput = 1 | 2 | 3 | 4 | 5 | 6 | 7;