+++ packages/office-ui-fabric-react/src/utilities/selectableOption/SelectableOption.types.ts
@@ -5,6 +5,11 @@ export interface ISelectableOption {
   key: string | number;
 
   /**
+   * ID attribute associated with this option
+   */
+  id?: string;
+
+  /**
    * Text to render for this option
    */
   text: string;