+++ packages/date-time/CHANGELOG.md
@@ -1,7 +1,19 @@
 # Change Log - @uifabric/date-time
 
-This log was last generated on Tue, 03 Dec 2019 12:36:28 GMT and should not be manually modified.
+This log was last generated on Wed, 11 Dec 2019 12:30:26 GMT and should not be manually modified.
 
+## 7.8.2
+Wed, 11 Dec 2019 12:30:26 GMT
+
+### Patches
+
+- Allow focus on CalendarMonth header. (phtucker@microsoft.com)
+## 7.8.1
+Fri, 06 Dec 2019 12:33:21 GMT
+
+### Patches
+
+- updated snapshots (betrue@microsoft.com)
 ## 7.8.0
 Tue, 03 Dec 2019 12:36:28 GMT